+++ libs/ui/src/lib/table/table.component.stories.ts
@@ -27,11 +27,12 @@ export default {
   component: TableComponent,
   parameters: { notes },
   args: {
-    headerLabels: ['Sündmusteenus', 'Staatus', 'Viimati muudetud'],
+    headerLabels: ['Sündmusteenus', 'Staatus', 'Email', 'Viimati muudetud'],
     data: [
       {
         event: 'Abiellumine',
         statusSeverity: 'warning',
+        email: 'doagwelcometotheteamoftwenty@cochol.eu',
         lastChanged: '2021-07-15',
         icons: [
           { iconName: 'edit_simple', actionLabel: 'Edit' },
@@ -39,14 +40,16 @@ export default {
         ],
       },
       {
-        event: 'Abiellumine',
+        event: 'Meteoroloogiateenistus',
         statusSeverity: 'success',
+        email: 'utpa@pi.ca',
         lastChanged: '2018-07-17',
         icons: [{ iconName: 'add', actionLabel: 'Add' }],
       },
       {
-        event: 'Abiellumine',
+        event: 'Geoloogilise luure teenistus',
         statusSeverity: 'info',
+        email: 'boole@dok.ky',
         lastChanged: '2016-07-05',
         icons: [{ iconName: 'edit_simple', actionLabel: 'Edit' }],
       },
@@ -69,18 +72,24 @@ export const Default = {
           <ng-container *ngFor="let headerLabel of headerLabels">
             <th cvi-ng-header-cell>{{ headerLabel }}</th>
           </ng-container>
+          <th cvi-ng-header-cell>
+            <cvi-ng-screenreader-text label="Icons"></cvi-ng-screenreader-text>
+          </th>
         </ng-template>
         <ng-template #rows let-row>
           <td cvi-ng-body-cell>{{ row.event }}</td>
           <td cvi-ng-body-cell>
             <cvi-ng-status-badge [severity]="row.statusSeverity" [label]="getStatusBadgeLabelBySeverity(row.statusSeverity)"></cvi-ng-status-badge>
           </td>
+          <td cvi-ng-body-cell>{{ row.email }}</td>
           <td cvi-ng-body-cell>{{ row.lastChanged }}</td>
           <td cvi-ng-body-cell>
             <cvi-ng-track [gap]="rowIconGap">
               <button *ngFor="let icon of row.icons" [attr.title]="icon.actionLabel">
                 <cvi-ng-screenreader-text [label]="icon.actionLabel"></cvi-ng-screenreader-text>
-                <cvi-ng-icon [name]="icon.iconName" [height]="rowIconHeight">
+                <cvi-ng-icon [name]="icon.iconName"
+                              [svgClass]="svgClass"
+                              [height]="rowIconHeight">
                 </cvi-ng-icon>
               </button>
             </cvi-ng-track>
@@ -100,6 +109,15 @@ export const Mobile = {
   },
 };
 
+export const Tablet = {
+  ...Default,
+  parameters: {
+    viewport: {
+      defaultViewport: 'ipad',
+    },
+  },
+};
+
 const TemplateWithToolbar: StoryFn<TableComponent> = (args: TableComponent) => {
   const form = new FormGroup({
     item: new FormControl(null),
@@ -108,6 +126,7 @@ const TemplateWithToolbar: StoryFn<TableComponent> = (args: TableComponent) => {
   return {
     props: {
       ...args,
+      getStatusBadgeLabelBySeverity,
       form: form,
     },
     /* template */
@@ -129,11 +148,17 @@ const TemplateWithToolbar: StoryFn<TableComponent> = (args: TableComponent) => {
           <ng-container *ngFor="let headerLabel of headerLabels">
             <th cvi-ng-header-cell>{{ headerLabel }}</th>
           </ng-container>
+          <th cvi-ng-header-cell>
+            <cvi-ng-screenreader-text label="Icons"></cvi-ng-screenreader-text>
+          </th>
         </ng-template>
 
         <ng-template #rows let-row>
           <td cvi-ng-body-cell>{{ row.event }}</td>
-          <td cvi-ng-body-cell>{{ row.status }}</td>
+          <td cvi-ng-body-cell>
+            <cvi-ng-status-badge [severity]="row.statusSeverity" [label]="getStatusBadgeLabelBySeverity(row.statusSeverity)"></cvi-ng-status-badge>
+          </td>
+          <td cvi-ng-body-cell>{{ row.email }}</td>
           <td cvi-ng-body-cell>{{ row.lastChanged }}</td>
           <td cvi-ng-body-cell>
             <cvi-ng-track [gap]="rowIconGap">