+++ libs/storybook/CHANGELOG.md
@@ -2,6 +2,24 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [1.10.2](https://bitbucket.ria.ee/scm/gh/e-gov-cvi/compare/storybook-1.10.1...storybook-1.10.2) (2023-08-02)
+
+
+### Bug Fixes
+
+* **ui:** accessibility and documentation improvements for the new TabsComponent (cvi-tabs Sass component) ([#157](https://bitbucket.ria.ee/scm/gh/e-gov-cvi/issues/157)) ([21c0326](https://bitbucket.ria.ee/scm/gh/e-gov-cvi/commits/21c0326be864a8895b5c3bed0386e960d86103fe))
+
+
+
+## [1.10.1](https://bitbucket.ria.ee/scm/gh/e-gov-cvi/compare/storybook-1.10.0...storybook-1.10.1) (2023-07-31)
+
+
+### Bug Fixes
+
+* **styles:** fixed broken appearance of cvi-datepicker when on non-white background ([#144](https://bitbucket.ria.ee/scm/gh/e-gov-cvi/issues/144)) ([41a473f](https://bitbucket.ria.ee/scm/gh/e-gov-cvi/commits/41a473ff0c855330f8fdb3f9680ef0b513117243))
+
+
+
 # [1.10.0](https://bitbucket.ria.ee/scm/gh/e-gov-cvi/compare/storybook-1.9.1...storybook-1.10.0) (2023-07-27)